feat!: do not embed token in another object
The access_token and token_type were being wrapped into another
TokenResponse object - this was breaking the Swagger docs authorize
button as it does not know where to get the access_token from.
The access_token and token_type are now top-level keys in the response
of /signup and /user/token.